Fix a bug related to the color pick option
The color pick tool option had a design flaw where it would not pick the white color of the initial squares rendered in the grid. However, if the user would use the white color, there would be no distinction between a "blank" square and a white one. So instead, the color pick tool is now picking up white from the blank squares.